Rename Announcer filenames to lowercase
